Fix for Empty Username
There was a problem if you left the username blank and tried to start
MC, it would close the launcher and nothing would happen.
Now it will only go ahead with the opening MC IF the username is not
blank, other wise it will throw an error with a doctor who refrence.