Reduced CPU Usage by atleast 40+ PCT + Other
So I noticed that the launcher was using like 40+ CPU all times when
launched, found out that BackgroundWorker1 was looping and doing nothing
since we don't use it. Now it sits at ~ 10% for a couple of seconds
(start up) then goes to <1% idle. Also deleted a few old things (server
status) and playing around with a animated BG for the menu. (looks bad
atm)